requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
cachetools
user-agents
pydantic-settings